pyproject.toml
[tool.poetry]
name = "spacestop-discord-bot"
version = "0.1.0"
description = "Discord Bot for viewing APOD"
authors = ["GermanIvk05 <48442292+GermanIvk05@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
